requirements.txt
APScheduler
arabic-reshaper
asgiref
asn1crypto
bs4
bleach
certifi
cffi
chardet
charset-normalizer
click
cryptography
cssselect2
Django==4.2.23
django-auditlog
django-apscheduler
# django-fernet-fields removed — incompatible with Django 4.x (uses removed force_text)
# Fernet encryption is implemented in horilla/encrypted_field.py using cryptography directly
django-cors-headers
django-environ
django-filter
django-haystack
django-import-export
django-jsonfield
django-mathfilters
django-microsoft-auth
django-model-utils
django-simple-history
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
drf-yasg
et-xmlfile
future
geographiclib
geopy
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
gunicorn
html5lib
Jinja2
idna
lxml
numpy
oauthlib
openpyxl
oscrypto
pandas
pdfkit
pillow
pycparser
pyHanko
pyhanko-certvalidator
PyMuPDF==1.24.5
pypdf
pypng
python-bidi
python-dateutil
pytz
PyYAML
pyzk==0.9
qrcode
reportlab==4.0.9
requests
responses
setuptools
simpleeval==0.9.13
six
sqlparse
svglib
swapper
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
utils
webencodings
Whoosh
xhtml2pdf
XlsxWriter
gunicorn
psycopg2-binary
whitenoise
django-storages
boto3
